requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
email-validator